requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.18
click==8.1.7
coloredlogs==15.0.1
cryptography==43.0.3
dataclasses-json==0.6.7
deprecated==1.2.14
duckduckgo-search==6.3.4
durationpy==0.9
emoji==2.14.0
eval-type-backport==0.2.0
fastapi==0.115.4
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
google-auth==2.36.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kubernetes==31.0.0
langchain==0.3.7
langchain-chroma==0.1.4
langchain-community==0.3.5
langchain-core==0.3.15
langchain-ollama==0.2.0
langchain-text-splitters==0.3.2
langdetect==1.0.9
langsmith==0.1.142
lxml==5.3.0
markdown-it-py==3.0.0
marshmallow==3.23.1
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
ollama==0.3.3
onnxruntime==1.20.0
opentelemetry-api==1.28.1
opentelemetry-exporter-otlp-proto-common==1.28.1
opentelemetry-exporter-otlp-proto-grpc==1.28.1
opentelemetry-instrumentation==0.49b1
opentelemetry-instrumentation-asgi==0.49b1
opentelemetry-instrumentation-fastapi==0.49b1
opentelemetry-proto==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
opentelemetry-util-http==0.49b1
orjson==3.10.11
overrides==7.7.0
packaging==24.2
posthog==3.7.0
primp==0.7.0
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.6.1
pygments==2.18.0
pypdf==5.1.0
pypika==0.48.9
pyproject-hooks==1.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-oxmsg==0.0.1
pyyaml==6.0.2
rapidfuzz==3.10.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlalchemy==2.0.35
starlette==0.41.2
sympy==1.13.3
tenacity==9.0.0
tokenizers==0.20.3
tqdm==4.67.0
typer==0.13.0
typing-extensions==4.12.2
typing-inspect==0.9.0
unstructured==0.16.5
unstructured-client==0.27.0
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.0
wrapt==1.16.0
yarl==1.17.1
zipp==3.21.0
